requirements.txt
openai>=1.99.9
pandas
playwright
python-dotenv
pydantic